type fixes

+++ audio_pattern_detector/audio_pattern_detector.py
@@ -626,7 +626,7 @@ class AudioPatternDetector:
     def _get_peak_times_normal(
         self,
         correlation_clip: NDArray[np.float64],
-        correlation_slice: NDArray[np.float32],
+        correlation_slice: NDArray[np.float64],
         seconds: list[float],
         peak: int,
         clip_name: str,
@@ -822,7 +822,7 @@ class AudioPatternDetector:
     def _get_peak_times_beep_v3(
         self,
         correlation_clip: NDArray[np.float64],
-        correlation_slice: NDArray[np.float32],
+        correlation_slice: NDArray[np.float64],
         seconds: list[float],
         peak: int,
         clip_name: str,
